Add check for having scenes when deleting camera

+++ src/main/java/org/eam/code/vmixapp/service/MyCameraService.java
@@ -3,14 +3,17 @@ package org.eam.code.vmixapp.service;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import org.eam.code.vmixapp.dao.MyCameraDAO;
+import org.eam.code.vmixapp.dao.SceneDao;
 import org.eam.code.vmixapp.model.MyCamera;
 import org.eam.code.vmixapp.model.Sequence;
 
 public class MyCameraService {
     private final MyCameraDAO cameraDAO;
+    private final SceneDao sceneDao;
 
     public MyCameraService() {
         this.cameraDAO = new MyCameraDAO();
+        this.sceneDao = new SceneDao();
     }
 
     public ObservableList<MyCamera> getCameras() {
@@ -26,10 +29,9 @@ public class MyCameraService {
     }
 
     public void deleteCam(int id) {
+        if (!sceneDao.getScenes().isEmpty()) {
+            throw new IllegalStateException("Camera has scene(s).");
+        }
         cameraDAO.deleteCam(id);
-    }
-
-    public MyCamera getCameraById(int id) {
-        return cameraDAO.getCameraById(id);
     }
 }
